Player done, all turrets done

+++ MyProject/Assets/Scripts/EnemyData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyData", menuName = "EnemyParameter", order = 1)]
 public class EnemyData : ScriptableObject
 {
-    public int hp;
-    public int atk;
+    public float hp;
+    public float atk;
     public float movespeed;
 }
