+++ MyProject/Assets/Scripts/TurretData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "TurretData", menuName = "TurretParameter", order = 1)]
 public class TurretData : ScriptableObject
 {
-    public int hp;
-    public int atk;
+    public float hp;
+    public float atk;
     public int cost;
     public float range;
     public float atkInterval;
